requirement.txt
absl-py==1.4.0
aiofiles==22.1.0
aiohttp==3.8.6
aiosignal==1.3.1
aiosqlite==0.19.0
alembic==1.12.1
amqp==5.1.1
annotated-types==0.6.0
antiorm==1.2.1
anyio==4.0.0
apache-airflow==2.7.3
apache-airflow-providers-celery==3.4.1
apache-airflow-providers-common-sql==1.8.0
apache-airflow-providers-ftp==3.6.0
apache-airflow-providers-http==4.6.0
apache-airflow-providers-imap==3.4.0
apache-airflow-providers-sqlite==3.5.0
apispec==6.3.0
appdirs==1.4.4
apturl==0.5.2
argcomplete==3.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
astroid==3.0.1
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
autopep8==2.0.4
Babel==2.13.1
backoff==1.10.0
bcrypt==3.2.0
beautifulsoup4==4.12.2
beniget==0.4.1
billiard==4.1.0
black==23.11.0
bleach==6.1.0
blinker==1.6.3
Brlapi==0.8.3
Brotli==1.0.9
bs4==0.0.1
cachelib==0.9.0
cachetools==5.3.2
cattrs==23.1.2
celery==5.3.4
certifi==2023.7.22
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
clickclick==20.10.2
cloudpickle==2.2.1
colorama==0.4.6
colorlog==4.8.0
comm==0.2.0
command-not-found==0.3
configobj==5.0.6
ConfigUpdater==3.1.1
connexion==2.14.2
contourpy==1.2.0
cron-descriptor==1.4.0
croniter==2.0.1
cryptography==41.0.5
cupshelpers==1.0
cycler==0.12.1
db==0.1.1
dbus-python==1.2.18
debugpy==1.8.0
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
Deprecated==1.2.14
deprecation==2.1.0
dill==0.3.1.1
distro==1.7.0
distro-info==1.1+ubuntu0.1
dnspython==2.4.2
docstring-parser==0.15
docstring-to-markdown==0.13
docutils==0.20.1
duplicity==0.8.21
elyra==3.15.0
elyra-examples-kfp-catalog==0.1.0
email-validator==1.3.1
entrypoints==0.4
exceptiongroup==1.1.3
executing==2.0.1
fasteners==0.14.1
fastjsonschema==2.19.0
findspark==2.0.1
fire==0.5.0
flake8==6.1.0
Flask==2.2.5
Flask-AppBuilder==4.3.6
Flask-Babel==2.0.0
Flask-Caching==2.1.0
Flask-JWT-Extended==4.5.3
Flask-Limiter==3.5.0
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
flower==2.0.1
fonttools==4.44.0
fqdn==1.5.1
frozenlist==1.4.0
fs==2.4.12
future==0.18.2
gast==0.5.2
GDAL==3.4.1
gitdb==4.0.11
GitPython==3.1.40
google-api-core==2.15.0
google-api-python-client==1.12.11
google-auth==2.25.1
google-auth-httplib2==0.1.1
google-cloud-core==2.4.1
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-re2==1.1
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
graphviz==0.20.1
greenlet==3.0.1
grpcio==1.59.2
gunicorn==21.2.0
h11==0.14.0
html5lib==1.1
httpcore==0.16.3
httplib2==0.20.2
httptools==0.6.1
httpx==0.23.3
humanize==4.8.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
inflection==0.5.1
ipykernel==6.27.1
ipython==8.18.1
ipython_genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
isort==5.12.0
itsdangerous==2.1.2
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter-resource-usage==1.0.1
jupyter-server-mathjax==0.2.6
jupyter-ydoc==0.2.5
jupyter_client==7.4.9
jupyter_core==5.5.0
jupyter_packaging==0.12.3
jupyter_server==2.11.1
jupyter_server_fileid==0.9.0
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.8.0
jupyterlab==3.6.6
jupyterlab-lsp==4.3.0
jupyterlab-widgets==3.0.9
jupyterlab_git==0.44.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
keyring==23.5.0
kfp==1.8.22
kfp-pipeline-spec==0.1.16
kfp-server-api==1.8.5
kfp-tekton==1.9.0
kfp-tekton-server-api==1.8.0rc8
kiwisolver==1.4.5
kombu==5.3.2
kubernetes==25.3.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy-object-proxy==1.9.0
limits==3.6.0
linkify-it-py==2.0.2
lockfile==0.12.2
louis==3.20.0
lxml==4.9.3
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.2.4
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
matplotlib==3.8.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mdit-py-plugins==0.4.0
mdurl==0.1.2
minio==7.2.0
mistune==3.0.2
monotonic==1.6
more-itertools==8.10.0
mpmath==0.0.0
multidict==6.0.4
mypy-extensions==1.0.0
mysql-connector-python==8.0.31
nbclassic==1.0.0
nbclient==0.9.0
nbconvert==7.11.0
nbdime==3.2.1
nbformat==5.9.2
nest-asyncio==1.5.8
netifaces==0.11.0
networkx==3.2.1
notebook==6.5.6
notebook_shim==0.2.3
numpy==1.26.1
oauthlib==3.2.0
olefile==0.46
opentelemetry-api==1.20.0
opentelemetry-exporter-otlp==1.20.0
opentelemetry-exporter-otlp-proto-common==1.20.0
opentelemetry-exporter-otlp-proto-grpc==1.20.0
opentelemetry-exporter-otlp-proto-http==1.20.0
opentelemetry-proto==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-semantic-conventions==0.41b0
ordered-set==4.1.0
orjson==3.9.10
overrides==7.4.0
OWSLib==0.25.0
packaging==23.2
pandas==2.1.2
pandocfilters==1.5.0
papermill==2.5.0
paramiko==2.9.3
parso==0.8.3
path==16.7.1
pathspec==0.11.2
pbr==5.8.0
pendulum==2.1.2
pexpect==4.8.0
Pillow==9.0.1
platformdirs==4.0.0
plotly==5.4.0
pluggy==1.3.0
ply==3.11
prison==0.2.1
prometheus-client==0.18.0
prompt-toolkit==3.0.41
protobuf==3.20.3
psutil==5.9.6
psycopg2==2.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.7
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycairo==1.20.1
pycodestyle==2.11.1
pycparser==2.21
pycryptodome==3.19.0
pycups==2.0.1
pydantic==1.10.13
pydantic-extra-types==2.1.0
pydantic-settings==2.1.0
pydantic_core==2.10.1
pydocstyle==6.3.0
pyflakes==3.1.0
PyGithub==2.1.1
Pygments==2.16.1
PyGObject==3.42.1
PyJWT==2.8.0
pylint==3.0.2
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyproj==3.3.0
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyRFC3339==1.1
pyrsistent==0.18.1
pyspark==3.5.0
python-apt==2.4.0+ubuntu2
python-daemon==3.0.1
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.0
python-gitlab==4.2.0
python-json-logger==2.0.7
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.9.0
python-multipart==0.0.6
python-nvd3==0.15.0
python-slugify==8.0.1
pythran==0.10.0
pytoolconfig==1.2.6
pytz==2023.3.post1
pytzdata==2020.1
pyxdg==0.27
PyYAML==6.0.1
pyzmq==24.0.1
QScintilla==2.11.6
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.30.2
reportlab==3.6.8
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.6.0
rich-argparse==1.4.0
rope==1.11.0
rpds-py==0.10.6
rsa==4.9
scipy==1.11.3
seaborn==0.13.0
SecretStorage==3.3.1
Send2Trash==1.8.2
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
snowballstemmer==2.2.0
soupsieve==2.5
spyne==2.14.0
SQLAlchemy==1.4.50
SQLAlchemy-JSONField==1.0.1.post0
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.4
stack-data==0.6.3
starlette==0.27.0
strip-hints==0.1.10
suds==1.1.2
sympy==1.9
systemd-python==234
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.3.0
terminado==0.18.0
terminator==2.1.1
text-unidecode==1.3
tinycss2==1.2.1
tomli==2.0.1
tomlkit==0.12.3
tornado==6.3.3
tqdm==4.66.1
traitlets==5.14.0
typer==0.9.0
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
uc-micro-py==1.0.2
ufoLib2==0.13.1
ufw==0.36.1
ujson==5.8.0
unattended-upgrades==0.1
unicodecsv==0.14.1
unicodedata2==14.0.0
uri-template==1.3.0
uritemplate==3.0.1
urllib3==1.26.18
usb-creator==0.3.7
uvicorn==0.24.0.post1
uvloop==0.19.0
vboxapi==1.0
vine==5.0.0
wadllib==1.3.6
watchdog==3.0.0
watchfiles==0.21.0
wcwidth==0.2.9
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
websockets==12.0
Werkzeug==2.2.3
whatthepatch==1.0.5
widgetsnbextension==4.0.9
wrapt==1.15.0
WTForms==3.0.1
xdg==5
xkit==0.0.0
y-py==0.6.2
yapf==0.40.2
yarl==1.9.2
yaspin==3.0.1
ypy-websocket==0.8.4
zipp==3.17.0
